src/requirements.txt
PyYAML==5.3.1
fastapi==0.78.0
uvicorn==0.17.6